+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "instances"]
-	path = instances
+	path = tests/data
 	url = https://github.com/ellmau/adf-instances.git
